figured this one out
took about 6 hours of dedicated effort to try different variations. Then I realized that the reason the 'timeLimitSpeed' wasn't considered a rejection was because it could only be rejected. I changed it to '(resolve, reject)' and voila! Now 'Promise.race()' understands that when 'timeLimitSpeed' finishes, reject is not the resolution, it is the rejection. and the output comes as intended. A subtle change that's important to remember for the future.

P.S. If it isn't clear yet, I'm working through the 30 days of Javascript problems on leetcode. I've been gone a couple weeks with a bunch of other busy work. Hopefully, I'll be back on my grind continuing to complete these.