today's daily leetcode problem
Not a difficult problem whatsoever, surprised it was a medium. Initially, I thought it was more complicated than it was as I thought the characters in the string had to match until they didn't which added a touch more in terms of complexity, still not much harder. Then I realized the characters have to exist in each string, with character order being maintained. This allowed me to remove almost all complexity entirely and what you see is as simple as it gets really.