todays daily leetcode problem
I thought this problem was asking me to create the palindrome using the given string, however, I then realized it was just asking to output the length of the would-be palindrome, which made it a lot easier. Also, I decided to play with hashmaps and learned to sort them by value which worked out well.