todays daily leetode problem (my solution sucks)
I found a solution to the problem but I know the solution could be better. I had an idea of how I could use hash maps again as a frequency map for the occurrence of each character in a word. The frequency map would store the least occurring frequency of each character across each word. Then transpose the frequency map to a string array and return it. I'll start a new folder called, "To Be Improved" and I'll put all the problems I'm dissatisfied with in there. And then put the improvement alongside each problem.