Improved!!
I'm back from a trip,  but I got the solution! I did the hashmap solution I previously spoke of before "The Big Folder Change" so unfortunately all those commits aren't saved. The complaint was that my previous solution sucked. I had a notion for how to improve it using hashmaps, and I was right! I worked on it for a couple of hours and boy, I've learned a lot about hashmaps and iterators in JS. They're awesome! I've effectively reduced the time complexity of the original solution from O(n^2*m) to O(n*m), and I'm proud of that. Additionally, I've made the code far more readable than it used to be thanks to all I learned about iterators. Hopefully, it's clear what is going on without having to look at those god-awful iterators I used to have.