an old problem i started once upon a time
I started this problem forever ago and came back to it a few times. I'm finally good enough to have solved it! The first struggle was returning the list as a ListNode. The next struggle was figuring out how to create the list in reverse. Originally, I turned the reversed numbers into regular numbers, added them, then reversed the result and converted it back into a list. Then remembering what I had learned I decided I would remove the conversion part entirely, and instead would perform the calculation from the given ListNodes. After a ton of struggle, I did it! However, I ended up in conditional hell, and decided this is something that'll be improved.