No more conditional hell
As you can see the conditionals were significantly reduced, with the help of ChatGPT. It was pretty smart, it reduced my first if/else statement by calling a conditional on l1, where as long as it wasn't null it got the value, it just made it 0. So much simpler than what I did, I'll be considering that in the future. In terms of addressing the carry, it reduced it simply by allowing the math to be the conditional itself. If the values were greater than 10 we divided the sum by 10 and rounded down to get 1, however, if the sum/10 resulted in less than 10 it would round down to 0, smart! It then simplified my creation of the list node if/else statement by immediately initiating the head (sentinel) and setting current as a pointer to it. it then slightly altered my final if statement to handle different cases, although for this problem the carry will always be 1 or 0. Amazing!