lucky stuff
I came across this problem the baby version of the last coding problem I solved. I used the code and the concepts I learned from the last one for this one. However, I struggled to figure out the if statements here and had to succumb to using chatGPT again. I'm cheating myself using chatGPT. I did all the variables essentially the same, and then it was just a matter of how to parse through the lists. I tried my way but just kept running into an infinite loop and couldn't understand why, so for the next problem I'm going to disallow any usage of chatGPT.  Then we'll really see if I'm learning.