decided to play coding
This is a terrible solution, but I'm happy about how quickly I came to it. It is a solution albeit. I'm debating improving this one or not. This is one of those situations where I know that improving it means either changing the entire approach to the solution or, trying to make this work without using hashmaps. I'll add it to the improvement list, but I'll work on it after I get more problems under my belt. I may change my mind tomorrow morning after some sleep. I'll be a bit more clear-headed and we'll see whether I choose to work on this or another problem.