yep i'm improving it
I want to improve it and today is the day OpenAI decided they wanted to work on ChatGPT because it "has issues". I'm not home for long, I want to do some coding problems. I'll have to spend more time away from this problem, and it may help me attack it better.