I DOUBTED MYSELF!
Yep... I doubted myself. Originally, when I thought about the approach to this problem I wanted to do it with a left and right pointer because it just made sense, I now know this is called a sliding window. I wasn't sure, however, of the purpose of the right pointer at the time, which led me to ditch the idea and create my first solution. I now know that the purpose of the right pointer is to track the current character/index and is used to calculate the current window size. Game changer, I'm glad I recognized that initializing a new map every time I found a dupe was massively space-wasting. I'm very happy with my solution. I used chatGPT with the consistent prompt "guide me like a teacher without showing me code" and this seems to be the best way for me to learn to code without cheating myself with an instantaneous solution.