yesterdays leetcode problem
This one was straightforward until I had to deal with the actual point of the problem. I realized I was wasting a lot of time doing every combination of numbers in the "squareSum" function. So I employed chatGPT to guide me on reducing this time waste. It told me I was wasting so much time because we don't have to do every single combination, what we do have to do is every combination resulting in less than c. This led me to an interesting incrementation method, and a rather ugly for loop. I began working on this briefly yesterday, until I saw that chatGPT was working again, and returned to the "longest substring without repeating characters" problem.