todays daily leetcode
I did the simplest solution. Then I checked the topics, which automatically sends this to the improve folder. One of the topics was binary search, I'm aware of binary search taking my code from a time complexity of O(n^2) to O(log n), a massive improvement. Only problem is, I'm not sure how to implement it because the arrays aren't sorted in any particular order. Maybe I don't understand how binary search works? As far as I was aware, it checks to see if your number is bigger or smaller than the chosen number and keeps cutting in half until you find your match. This means I have to find a way to start cutting my search in half by first setting the difficulty and profit as key, value pairs in a different array, ordering them by difficulty. Then I would do binary search until we find all of the difficulties that are <= worker difficulty, and find the highest profit of the bunch. Is that really the solution though? I'm not sure, I'll probably revisit this at a later date because I'm doing some work for the next few days.