learning to toy with algorithms
A small alteration to the classic binary search. I have been watching more YouTube videos about coding and I'm starting to see bitwise versions of certain algorithms that would make things faster. I do wonder if doing bitwise operations for this would be faster or the same, but I wouldn't know how to approach that. I don't feel the need to improve this.