30 days of Javascript is really hard
I wouldn't know how to do any of this without looking it up. That being said, the way this works is that it is taking the function and its arguments as well as a time when that is supposed to happen. Then debounce is called at a certain time. If the time at which debounce is called is before a function is about to happen, we cut off that function and reset the timer, we continue to do that until we reach the final call of debounce. Sometimes, the function will be processed before the next call of debounce. The use cases could be when things are called multiple times in rapid succession, like spam prevention.