recursive stuff
Pretty simple recursion stuff, but there is a challenge asking to do this exact thing iteratively. I assume this will improve on the memory side of things due to the recursive function calls happening. I'm not sure how I would do this but I would like to try to do it from a while loop because I think that would be the best way for me to go, but still something feels wrong about that.