This isn't even correct
I just want to show that I'm working on a solution to this problem. It isn't technically incorrect but this is for too slow as it's O(nk), "k" being the window size. I'm looking at the topics and hints and they're suggesting queue, heap (priority queue), and deque. So I'll be learning what a deque is. And that'll get me to do this in O(n) time instead of O(nk), which should be a massive improvement.