Let's FREAKING GO
So I found the O(n) solution and I used an array as the monotonic deque, which is pretty cool. I noticed that my solution was on the lower end of both graphs on Leetcode. Normally, that means I'm missing something that everybody else is doing and I would go and find that better solution, but this is my first hard problem and I'm genuinely happy that I found it as fast as I did. Fantastic progress and my algorithm knowledge is growing, very excited!