I lied lolsies
I had to go check everyone else's solutions and then I saw that there was one that was the same as mine, EXCEPT theirs simplified my if statement into for loop. I guess I should've seen that coming because it's that situation I took notes on where the condition exists outside of the loop and inside the loop. The conditional itself becomes redundant, so what this does is it moves the conditional to a while loop that performs the if statement until it can't which is great! That's a good way of thinking of while loops, and maybe loops in general, that being "iterative if statements". Should've been something I was already doing, but with practice it'll solidify more. I'm just a lil jelly doin his best to become diamond, what can I say \/(^-^)\/