using my big JS brain
slammed 2 if statements into 1 line just because why the heck not, it makes it that little bit faster. I used the ternary operator just to not have to have the 2 and it's that much better. Also, I find it interesting that when I submit this in leetcode it is either high percentage time or memory, but not both. Maybe that's just my luck, it's probably 90%+ on both, which makes me feel awesome.