started leetcode 75
decided to start doing the interview questions, because it's about time I start interviewing. This question I took a peek at the solutions once I made mine and realized I was doing a lot of unnecessary calculations, and I was using a while loop and being dumb. This works great and is nice and simple with no conditionals other than what is in the for loop.