more sliding window
I had to look at the discussion board to find out it was a sliding window problem. Originally I had done it with the double for loop which is too slow. It was one of those situations where I knew it could be done in a single for loop and I didn't know how. What's a little annoying is that I knew the technique, but I didn't think to use it. I should list problem-solving techniques frequently used to refer to so I can do it myself.