bit manipulation?!
Bitwise operations are still crazy to me, but this is apparently the only way this can be solved. Very cool to see the ideation behind this problem. I had to look at the discussion again to see the thought process for this, and I'm not upset I didn't know how to do this, because this is crazy. Essentially all these numbers are being treated as binary numbers and being Xor'ed (1s = 0, 1 = 1, and 0 = 0). Then magic and kaboom right answer.