set-c solution ;)
I did this with double for loop then time limit exceeded. Then I did this with a map because it worked with a single loop. Then I remembered sets exist which is great because they remove duplicates. This allowed me to remove the loop and the if statement to catch the dupe entirely. This left me with initializing a set and checking its size compared to the array length. Then I remembered I could implicitly initialize the set and directly return the conditional.