todays daily leetcode
Todays daily happened to line up with what I was training anyway, which is array problems. My solution is fine. I expected the logs to have more characters and variation, which is why I approached it with the "charAt" function. I thought that there would be characters when moving up and down the directories, which I guess wasn't the case. Either way, we can consider it my personal edge case or something. It was completed first try which felt nice.