perf: Improve numToEng function performance by 50% through optimized logic and reduced string operations
Found out how to remove the for loop and make better conditionals. Once I figured that out I removed "num.toString().split('')" as those tend to be pretty slow. Started dealing with the number directly again to get the individual digits. Then, I removed the strings in the hashmap because integers are faster than strings. Used the ternary operator to better handle adding spaces, which was really a minor issue in the previous iteration.