chore: add AoC problems and remove license from Numbers to English
I want to tackle Advent of Code problems for the first time because it looks fun. I felt like the licensing from the old repo was unnecessary so just doing some cleaning up. This is mostly just prepping to do a bunch of practice problems. I have a lot of coursework to do so we'll see when I get around to actually doing these problems.