Add: leetcode solution for problem 438
Ran into tons of issues trying to solve this problem. At the beginning I was looking at every possible permutation of p which was a horrible idea, and although only needed to be done once, with a long enough string is so complex that it's completely dumb. Then I finally remembered frequency maps, and I made it work with that in conjunction with a sliding window approach. Except I was wasting a ton of space creating frequency maps for every valid substring. So then, I created one frequency map populating it only with the important characters, only updating the values when there is a valid substring. This approach is straightforward and intuitive which I like but can be greatly improved. For one, the entire frequency map needs to be set to 0 after each valid substring is checked. For two, the sliding window is changing sizes and that's actually hurting my efficiency because I need the window to remain constant because I'm adding extra iterations when it isn't. This one is definitely going to the To Be Improved folder.