Add: twoSumLC.js solution
I wish I could've come up with the hashmap solution it's so smart. At the very least, I understand it directly, I'll call it two sides of the same story. Pros of mine, space complexity of constant time. But honestly, that's really not a pro given just how much space isn't an issue. In addition to the fact that the other solution takes the time complexity down to O(n) which is way way way worth increasing the space complexity to O(n). Great learning moment.