Add: groupAnagramsLC.js solution
I left the notes and ideas commented in the code to demonstrate my thought process. The main challenges I faced here were figuring out what I could use as a valid key/value for the map. Originally, I was using a map as the key, which turned out to be a bad idea because maps with the same content aren't considered the same due to being stored as references to an object. This means that the pointer to where each of these maps are being stored is what is being stored in the map as opposed to the content itself. So I changed it to the ordered string because strings are not an issue. Then my smaller issue was that I wanted to essentially do what I do for frequency maps and retrieve the value and if it didn't exist retrieve the most basic thing of the same type and then add to it. As it turns out, when you do that for arrays you're actually getting the length of the array and pushing that instead of the actual updated array. This meant I had to break it apart and separately set/get the array in order to properly update it.