Add: binaryTreeLevelOrderTraversalLC.js solution
I didn't know how to do this. I rubber-ducked, wrote things down, and tried to imagine how I would approach this problem. Sadly, I just couldn't figure it out. So I had to look up how to do BFS (Breadth-First Search), which brought me to a queue. This is great because it helped me recall the idea of a FIFO data structure (queue) necessary to make this happen. It feels so intuitive yet so strange at the same time, hard to explain. I took a minute to write everything down and even visually mapped it for myself and how it is enqueued and dequeued to achieve BFS. This then brought me to the next set of problems. The first was understanding how to access nodes and values without hitting null errors, which I have commented in the code. The next issue was figuring out how I was going to get to each level of the array. However, while asking ChatGPT to help me understand an error I was encountering with accessing nodes, it flashed me the answer for absolutely no reason. This is one of the most frustrating things. I just couldn't unsee it, I saw it for a split second at most, but the damage was done. I was able to recall the basic idea for how to tackle the rest of the problem elegantly from just that glance. Disappointing as it may be, it was much better than the approach I was planning on doing. My approach involved creating a lower bound (2^n) and upper bound (2^(n+1)). Then, I track every single attempt at adding a node, and then when I reach the upper bound, I know I'm at the next level, allowing me to increase the lower and upper bound by 1 and continue. If you can't tell, I was getting rather complicated. The ChatGPT solution allowed me to see the rather simple solution, which was seeing what was currently enqueued and how much is on the current level and dequing that while adding the children, then on the next iteration, you then check queue size again and BOOM, that's the new next level. It's a way easier approach.