Add: binaryTreeRightSideViewLC.js solution
An incredible feeling to come up with a solution that beats 100% on my first submission. I understood the problem immediately, thanks to yesterday's problem. The thought process was so short that it was logged in all the comments. I don't have to explain much, either. I'm proud of the speed by which I found the solution, the recognition of the type of solution, and the recall of a past solution without needing to revisit. Self high five!