Add: pathSumLC.js solution
I specifically tasked myself to complete this problem using a recursive DFS structure. Otherwise, it'd be much easier to solve iteratively. In the solution, I've included the optimized solution to show the difference between what I did and what the optimal solution did differently. Namely, it removed unnecessary variables and got straight to the point, returning true or false exactly when it should. Mine did the same except using bad practices, which I was aware of while I was doing it. The problem was that I wasn't sure how to alter the code while maintaining best practices and the solution. Some bad practices I was using, "result" being a variable outside of the scope of dfs. This is bad because it introduces side effects that make the function harder to debug. At one point, I came close to fixing it by using return values, but I couldn't figure out how to make dfs return true or false, giving up shortly after.  The next bad practice was not returning recursive calls (this goes along with the first problem). If the goal was to return true or false from the recursive call, the recursive function should have been returning values from the start. And lastly (again due to the first bad practice), result == true in the base case to return based on an external variable instead of having the function return true or false.