Add: lowestCommonAncestorOfABinaryTreeLC.js solution
This took a long time because I decided I was going to do everything I could not to look any answers up. So what you see is the result of questioning my thinking and understanding and battling against the satisfaction of coming up with a valid solution. My first idea is included below the final solution to demonstrate where I started versus where I ended. My final solution was quite the battle to reach because I had to question my understanding of the question at every level. Is it necessary to store the paths to the nodes? Do I have to have an object that tells me whether or not I found the nodes I'm looking for? Is Pre-order traversal the best way to do this? Can I solve it by directly passing information up the recursive stack? These are some of the questions I had to ask myself to reach my final solution. It took me 15 iterations to get there, trying wrong answer after wrong answer. I feel a step closer to turning my mind into that of a programmer, one who is willing to fail over and over again to find an efficient solution. I'm happy about this.